+++ src/main/java/de/samply/converter/Converter.java
@@ -1,8 +1,8 @@
 package de.samply.converter;
 
-import de.samply.result.container.template.ResultTemplate;
+import de.samply.template.conversion.ConversionTemplate;
 import reactor.core.publisher.Flux;
 
 public interface Converter<I,O> {
-  Flux<O> convert (Flux<I> input, ResultTemplate template);
+  Flux<O> convert (Flux<I> input, ConversionTemplate template);
 }