+++ src/main/java/de/samply/converter/ConverterGroup.java
@@ -1,6 +1,6 @@
 package de.samply.converter;
 
-import de.samply.result.container.template.ResultTemplate;
+import de.samply.template.conversion.ConversionTemplate;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicReference;
 import reactor.core.publisher.Flux;
@@ -13,7 +13,7 @@ public class ConverterGroup<I,O> implements Converter<I,O>{
   }
 
   @Override
-  public Flux<O> convert(Flux<I> input, ResultTemplate template) {
+  public Flux<O> convert(Flux<I> input, ConversionTemplate template) {
     AtomicReference<Flux> tempFlux = new AtomicReference<>(input);
     converters.forEach(converter -> tempFlux.set(converter.convert(tempFlux.get(), template)));
     return tempFlux.get();