+++ src/main/java/de/samply/ldm/LdmClient.java
@@ -1,28 +1,21 @@
 package de.samply.ldm;
 
-import de.samply.converter.Format;
-import de.samply.query.Query;
-import de.samply.query.QueryManager;
-import de.samply.result.container.template.ResultTemplateManager;
-import de.samply.source.Source;
-import de.samply.source.SourcesManager;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 @Component
 public class LdmClient {
-
-  private final SourcesManager sourcesManager;
-  private final ResultTemplateManager resultTemplateManager;
+/*
+  private final ConverterTemplateManager converterTemplateManager;
+  private final ConversionTemplateManager conversionTemplateManager;
   private final QueryManager queryManager;
 
   public LdmClient(
-      @Autowired SourcesManager sourcesManager,
+      @Autowired ConverterTemplateManager converterTemplateManager,
       @Autowired QueryManager queryManager,
-      @Autowired ResultTemplateManager resultTemplateManager) {
-    this.sourcesManager = sourcesManager;
+      @Autowired ConversionTemplateManager conversionTemplateManager) {
+    this.conversionTemplateManager = converterTemplateManager;
     this.queryManager = queryManager;
-    this.resultTemplateManager = resultTemplateManager;
+    this.conversionTemplateManager = conversionTemplateManager;
   }
 
   public String retrieve(String sourceId, String queryId, String resultTemplateId,
@@ -37,17 +30,17 @@ public class LdmClient {
 
   public String retrieve(String sourceId, String query, Format queryFormat,
       String resultTemplateId, Format resultFormat) throws LdmClientException {
-    Source source = sourcesManager.getSource(sourceId);
-    if (source == null) {
+    ConverterTemplate converterTemplate = conversionTemplateManager.getSource(sourceId);
+    if (converterTemplate == null) {
       throw new LdmClientException("Source " + sourceId + " not found");
     }
-    if (!source.getQueryFormats().contains(queryFormat)) {
+    if (!converterTemplate.getQueryFormats().contains(queryFormat)) {
       throw new LdmClientException(
           "Query Format " + queryFormat + " not defined for source " + sourceId);
     }
     //TODO
     return null;
   }
-
+*/
 
 }