+++ src/test/java/de/samply/converter/ConverterManagerTest.java
@@ -1,9 +1,7 @@
 package de.samply.converter;
 
-import static org.junit.jupiter.api.Assertions.*;
-
-import de.samply.result.container.template.ResultTemplate;
-import de.samply.result.container.template.ResultTemplateManager;
+import de.samply.template.conversion.ConversionTemplate;
+import de.samply.template.conversion.ConversionTemplateManager;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import reactor.core.publisher.Flux;
@@ -15,9 +13,9 @@ class ConverterManagerTest {
   void getConverter() {
     ConverterManager converterManager = new ConverterManager();
     Converter converter = converterManager.getConverter(Format.FHIR_QUERY, Format.CSV);
-    ResultTemplateManager resultTemplateManager = new ResultTemplateManager("./templates");
-    ResultTemplate resultTemplate = resultTemplateManager.getResultTemplate("test-template1");
-    Flux flux = converter.convert(Flux.just("Patient"), resultTemplate);
+    ConversionTemplateManager conversionTemplateManager = new ConversionTemplateManager("./templates");
+    ConversionTemplate conversionTemplate = conversionTemplateManager.getConversionTemplate("test-template1");
+    Flux flux = converter.convert(Flux.just("Patient"), conversionTemplate);
     flux.blockLast();
     //TODO
   }